requirements.txt
Pillow
pytesseract
PyMuPDF
flask